[SDCP-536] French News Events List
