+++ server/templates/french_news_events_list_export.html
@@ -1,0 +1,11 @@
+{% for date, date_group in group_items_by_french_topics(items) %}
+<h2>{{ date_group.name }}</h2>
+{% for group, items in date_group.groups.items() %}
+{% if items %}
+<p><b>{{ group }}</b></p>
+{% for item in items %}
+<p>{{ item.title }}<br>{{ item.address.name }} _ {{ item.description }}{{ item.local_time }}{{ item.address.full }}{{ item.coverage_types }}</p>
+{% endfor %}
+{% endif %}
+{% endfor %}
+{% endfor %}
